pyproject.toml
[project]
name = "py-dagster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagster",
    "dagster-webserver",
    "python-dotenv",
    "pandas",
    "sqlalchemy",
    "pyodbc",
]

[tool.uv]